requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.0
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
watchdog==3.0.0